fix(Button): Fix conflict with HTML type
There are only three permissible types for an html button - 'submit',
'reset' and 'button'. Any other type is treated as a 'submit' type. This
leads to unexpected behavior when using a non-submit button inside a
form (Hitting enter triggers the onClick handler of the non-submit
button). The correct behavior would be to set the type to 'button' in
this case and to 'submit' only when the button is intended as the
primary button of a form.
The button's type can now be set independently of the button's
canonStyle.

This fixes #81