feat(FormPopover): Add additionalFooterContent prop
Add a new prop to the FormPopover component in order to allow developers
to display additional right-aligned content with the control
buttons/links (e.g. submit and cancel).

This feature is taking is consideration the vast majority of popover
examples we've seen so far, where people customize the footer by pulling
tooltips to the right. Content on the "right" of a footer strongly imply
the entire popover context and cleanly separates itself from the form
controls. Pulling to the right also plays nicely with the processing
indicator on submittable popovers.

This is a backwards compatible change and shouldn't require users to
anything on their side if they are not using the new prop.